Moved some gradle-independent classes to libxcode. Plugin jar includes all classes from libxcode, so it should not be changed.
