+++ libxcode/build.gradle
@@ -1,0 +1,23 @@
+apply plugin: 'groovy'
+
+repositories {
+	mavenCentral()
+}
+
+dependencies {
+	compile localGroovy()
+	compile 'commons-io:commons-io:2.2'
+	compile 'commons-configuration:commons-configuration:1.10'
+	compile 'commons-collections:commons-collections:3.2.2'
+	compile 'ch.qos.logback:logback-core:1.0.13'
+	compile 'org.slf4j:slf4j-api:1.7.21'
+
+	testCompile 'junit:junit:4.12'
+	testCompile 'org.hamcrest:hamcrest-all:1.3'
+	testCompile 'org.codehaus.groovy:groovy-all:2.4.4'
+	testCompile 'org.objenesis:objenesis:1.2'
+	testCompile 'cglib:cglib-nodep:2.2.2'
+	testCompile('org.spockframework:spock-core:1.0-groovy-2.4') {
+		exclude module: 'groovy-all'
+	}
+}
