+++ libxcode/src/main/groovy/org/openbakery/Version.groovy
