+++ libxcode/src/main/groovy/org/openbakery/simulators/SimulatorDevicePair.groovy
