+++ libxcode/src/main/groovy/org/openbakery/simulators/SimulatorDeviceType.groovy
