+++ libxcode/src/main/groovy/org/openbakery/util/StringHelper.groovy
