+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'xcode-plugin'
+
+include 'libxcode', 'plugin'
+
